add FTP container

+++ srcs/requirements/wordpress/tools/wordpress.sh
@@ -1,22 +1,25 @@
 #!/bin/bash
 
-mkdir -p /var/www/html
+HTML_FOLDER="/var/www/html"
 
-sed -i "s/listen = \/run\/php\/php7.4-fpm.sock/listen = 9000/" /etc/php/7.4/fpm/pool.d/www.conf
-cd /var/www/html
-rm -rf *
-sleep 10
+if [ -d "$HTML_FOLDER" ]; then
+    echo "$HTML_FOLDER The folder exists."
+else
+    mkdir -p $HTML_FOLDER
 
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar 
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/local/bin/wp
+    sed -i "s/listen = \/run\/php\/php7.4-fpm.sock/listen = 9000/" /etc/php/7.4/fpm/pool.d/www.conf
+    cd $HTML_FOLDER
+    rm -rf *
+    sleep 10
 
-wp core download --allow-root
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar 
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
 
-
-
-wp config create --dbname=$DATABASE_NAME --dbuser=$DB_USER --dbpass=$DB_PASSWORD --dbhost=$DB_HOST --allow-root
-wp core install --url=$DOMAINE_NAME --title=$WP_TITLE --admin_user=$WP_ADMIN_USER --admin_email=$WP_ADMIN_EMAIL --admin_password=$WP_ADMIN_PASSWORD --skip-email --allow-root
-wp user create $WP_USER $WP_EMAIL --role=author --user_pass=$WP_PASSWORD --allow-root
+    wp core download --allow-root
+    wp config create --dbname=$DATABASE_NAME --dbuser=$DB_USER --dbpass=$DB_PASSWORD --dbhost=$DB_HOST --allow-root
+    wp core install --url=$DOMAINE_NAME --title=$WP_TITLE --admin_user=$WP_ADMIN_USER --admin_email=$WP_ADMIN_EMAIL --admin_password=$WP_ADMIN_PASSWORD --skip-email --allow-root
+    wp user create $WP_USER $WP_EMAIL --role=author --user_pass=$WP_PASSWORD --allow-root
+fi
 
 exec $@
